Implement an ugly hack in the zephyr sx127x drivers. It hat checks every 10 seconds if the Radio is connected during RX. This prevents the radio from going silent without an error if a voltage drop made it go offline or similar.
Also adds a check during driver init that prevents it from
executing further if it gets a bad version number.
Using our boards the SPI dosen't error if it can't reach
the sx1276 but gives a bad version number.
If it can't reach the Radio it stops the entire programs execution after
the version check. This fix prevents that